+++ spec/integration/http_spec.rb
@@ -6,7 +6,7 @@ describe 'Http rack setup' do
   before { app.wait_start }
 
   context 'When booting' do
-    let(:middleware) { app_helper.shell_run "cd #{app_path} && rake middleware" }
+    let(:middleware) { app_helper.shell_run "cd #{app_path} && rails middleware" }
 
     it 'inserts rack timeout into the middleware stack' do
       expect(middleware).to include 'Rack::Timeout'